refactor: 리뷰 모달 컴포넌트 폴더 구조 개선 및 import 경로 수정
