+++ apps/web/app/review-modal/[reviewId]/search/video/page.client.tsx
@@ -11,7 +11,10 @@ import {
   ApiReviewSearchResponse,
   VideoReviewResponse,
 } from 'app/api/review-response';
-import ReviewOnboardingModal from 'app/review-modal/components/ReviewOnboardingModal';
+import {
+  ReviewModalSwiper,
+  ReviewOnboardingModal,
+} from 'app/review-modal/components';
 import LoadingSvg from 'components/loading/loading-svg';
 import { REVIEW_KEYS } from 'constants/query-key';
 import type {
@@ -27,7 +30,6 @@ import {
   isValidCategoryKey,
   isValidCategoryOption,
 } from '../../../../../utils/category';
-import ReviewModalSwiper from '../../../components/review-modal-swiper';
 import { useAllVideoReviewDetails } from '../../../hooks/review-api';
 import type { ReviewDetail } from '../../../types';
 
@@ -88,6 +90,8 @@ export default function ClientPage({ userStatus }: VideoReviewClientPageProps) {
   const handleCloseOnboarding = () => {
     setIsOnboardingOpen(false);
   };
+
+  const detailQueries = useAllVideoReviewDetails(reviewData);
   const currentIndex = reviewData.findIndex(
     (review) => review.reviewId === currentReviewId
   );
@@ -98,8 +102,11 @@ export default function ClientPage({ userStatus }: VideoReviewClientPageProps) {
     }
   }, [currentIndex, router]);
 
-  const detailQueries = useAllVideoReviewDetails(reviewData);
+  if (currentIndex === -1) {
+    return null;
+  }
 
+  // 모든 상세 정보가 로딩 완료될 때까지 대기
   if (detailQueries.some((q) => q.isLoading)) {
     return (
       <div className="flex min-h-screen w-full items-center justify-center bg-black">
@@ -108,6 +115,7 @@ export default function ClientPage({ userStatus }: VideoReviewClientPageProps) {
     );
   }
 
+  // 상세 정보를 ID로 매핑
   const detailMap = new Map<number, VideoReviewDetailResponse>();
   reviewData.forEach((review, index) => {
     const dq = detailQueries[index];
@@ -119,6 +127,7 @@ export default function ClientPage({ userStatus }: VideoReviewClientPageProps) {
     }
   });
 
+  // 슬라이더에 넘길 리뷰 데이터 구성
   const allReviews: ReviewDetail[] = reviewData
     .filter((review) => {
       const detail = detailMap.get(review.reviewId);
@@ -142,7 +151,7 @@ export default function ClientPage({ userStatus }: VideoReviewClientPageProps) {
         brandName: detail.brandName,
         productName: detail.productName,
         productImageUrl: detail.productImageUrl,
-        mediaList: detail.videoUrls.map((url: string, index: number) => ({
+        mediaList: detail.videoUrls.map((url, index) => ({
           id: index,
           type: 'video' as const,
           url,
