+++ apps/web/app/review-modal/components/index.ts
@@ -1,0 +1,3 @@
+export * from './media';
+export * from './review';
+export * from './modal-structure';
