+++ apps/web/app/review-modal/components/media/media-info.tsx
@@ -6,7 +6,7 @@ import { Avatar } from '@lococo/design-system/avatar';
 import { ReactionToggle } from '@lococo/design-system/reaction-toggle';
 import { SvgGoodFill, SvgGoodOutline, SvgSend } from '@lococo/icons';
 
-import { useReviewLikeToggle } from '../hooks/review-api';
+import { useReviewLikeToggle } from '../../hooks/review-api';
 
 interface MediaInfoProps {
   reviewId: number;
