+++ apps/web/app/review-modal/components/modal-structure/index.ts
@@ -1,0 +1,3 @@
+export { default as ReviewModalSwiper } from './review-modal-swiper';
+export { default as ReviewModalLayout } from './review-modal-layout';
+export { default as ReviewOnboardingModal } from './ReviewOnboardingModal';
