+++ apps/web/app/review-modal/components/review/index.ts
@@ -1,0 +1,2 @@
+export { default as ReviewInfo } from './review-info';
+export { default as Comment } from './comment';
