+++ apps/web/app/review-modal/components/review/review-info.tsx
@@ -6,8 +6,8 @@ import { Star } from '@lococo/design-system/star';
 import { Tag } from '@lococo/design-system/tag';
 import { SvgArrowOutward, SvgClose } from '@lococo/icons';
 
-import { ReviewDetail } from '../types';
-import Comment from './comment';
+import { ReviewDetail } from '../../types';
+import { Comment } from './';
 
 type ReviewInfoProps = Pick<
   ReviewDetail,
